% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{gdax_api}
\alias{gdax_api}
\alias{gdax_get}
\alias{gdax_post}
\alias{gdax_put}
\alias{gdax_delete}
\title{GDAX}
\usage{
gdax_api(endpoint, method = c("get", "post", "put", "delete"), ...)

gdax_get(endpoint, ...)

gdax_post(endpoint, ...)

gdax_put(endpoint, ...)

gdax_delete(endpoint, ...)
}
\arguments{
\item{method}{appropriate web method. These map to httr::GET, httr::POST, etc...}

\item{...}{extra arguments passed to httr::GET, httr::POST, etc...}
}
\description{
Make request to GDAX
}
\examples{
\dontrun{
gdax_get("accounts", query = list(date = "2017-01-01"))
}
}
\keyword{internal}
