% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.r
\name{get_account}
\alias{get_account}
\title{Get an Account}
\usage{
get_account(account_id)
}
\arguments{
\item{account_id}{Account Id}
}
\description{
Information for a single account. Use this endpoint when you know the account_id.
}
\references{
\url{https://docs.gdax.com/#get-an-account}
}
\seealso{
Other accounts: \code{\link{get_account_history}},
  \code{\link{get_account_holds}},
  \code{\link{get_accounts}}
}
