% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.r
\name{get_account_history}
\alias{get_account_history}
\title{Get Account History}
\usage{
get_account_history(account_id)
}
\arguments{
\item{account_id}{Account Id}
}
\description{
List account activity. Account activity either increases or decreases your
account balance. Items are paginated and sorted latest first. See the
Pagination section for retrieving additional entries after the first page.
}
\references{
\url{https://docs.gdax.com/#get-account-history}
}
\seealso{
Other accounts: \code{\link{get_account_holds}},
  \code{\link{get_accounts}}, \code{\link{get_account}}
}
