% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.r
\name{get_account_holds}
\alias{get_account_holds}
\title{Get Holds}
\usage{
get_account_holds(account_id)
}
\arguments{
\item{account_id}{Account Id}
}
\description{
Holds are placed on an account for any active orders or pending withdraw
requests. As an order is filled, the hold amount is updated. If an order is
canceled, any remaining hold is removed. For a withdraw, once it is
completed, the hold is removed.
}
\references{
\url{https://docs.gdax.com/#get-account-history}
}
\seealso{
Other accounts: \code{\link{get_account_history}},
  \code{\link{get_accounts}}, \code{\link{get_account}}
}
