% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.r
\name{get_accounts}
\alias{get_accounts}
\title{List Accounts}
\usage{
get_accounts()
}
\description{
Get a list of trading accounts.
}
\references{
\url{https://docs.gdax.com/#list-accounts}
}
\seealso{
Other accounts: \code{\link{get_account_history}},
  \code{\link{get_account_holds}},
  \code{\link{get_account}}
}
